Handle CAST() and CONVERT() functions - fixes #25
Move get_query_columns tests to a separate file
